ReactComponentのレファクタリング
